Move validator to *_validator.go. Changes with validator "transition" and "end" validator: add to BaseState; fix failed unit tests; refactor SwitchState validator.
