+++ model/parallel_state.go
@@ -15,16 +15,10 @@
 package model
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
-	"reflect"
-	"strconv"
 
-	validator "github.com/go-playground/validator/v10"
 	"k8s.io/apimachinery/pkg/util/intstr"
-
-	val "github.com/serverlessworkflow/sdk-go/v2/validator"
 )
 
 // CompletionType define on how to complete branch execution.
@@ -108,37 +102,4 @@ type BranchTimeouts struct {
 type ParallelStateTimeout struct {
 	StateExecTimeout  *StateExecTimeout `json:"stateExecTimeout,omitempty"`
 	BranchExecTimeout string            `json:"branchExecTimeout,omitempty" validate:"omitempty,iso8601duration"`
-}
-
-// ParallelStateStructLevelValidation custom validator for ParallelState
-func ParallelStateStructLevelValidation(_ context.Context, structLevel validator.StructLevel) {
-	parallelStateObj := structLevel.Current().Interface().(ParallelState)
-
-	if parallelStateObj.CompletionType == CompletionTypeAllOf {
-		return
-	}
-
-	switch parallelStateObj.NumCompleted.Type {
-	case intstr.Int:
-		if parallelStateObj.NumCompleted.IntVal <= 0 {
-			structLevel.ReportError(reflect.ValueOf(parallelStateObj.NumCompleted), "NumCompleted", "numCompleted", "gt0", "")
-		}
-	case intstr.String:
-		v, err := strconv.Atoi(parallelStateObj.NumCompleted.StrVal)
-		if err != nil {
-			structLevel.ReportError(reflect.ValueOf(parallelStateObj.NumCompleted), "NumCompleted", "numCompleted", "gt0", err.Error())
-			return
-		}
-
-		if v <= 0 {
-			structLevel.ReportError(reflect.ValueOf(parallelStateObj.NumCompleted), "NumCompleted", "numCompleted", "gt0", "")
-		}
-	}
-}
-
-func init() {
-	val.GetValidator().RegisterStructValidationCtx(
-		ParallelStateStructLevelValidation,
-		ParallelState{},
-	)
 }
