+++ model/parallel_state_test.go
@@ -20,8 +20,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/util/intstr"
-
-	val "github.com/serverlessworkflow/sdk-go/v2/validator"
 )
 
 func TestParallelStateUnmarshalJSON(t *testing.T) {
@@ -64,138 +62,6 @@ func TestParallelStateUnmarshalJSON(t *testing.T) {
 
 			assert.NoError(t, err)
 			assert.Equal(t, tc.expect, &v)
-		})
-	}
-}
-
-func TestParallelStateStructLevelValidation(t *testing.T) {
-	type testCase struct {
-		desp  string
-		state *State
-		err   string
-	}
-	testCases := []testCase{
-		{
-			desp: "normal",
-			state: &State{
-				BaseState: BaseState{
-					Name: "1",
-					Type: "parallel",
-				},
-				ParallelState: &ParallelState{
-					Branches: []Branch{
-						{
-							Name: "b1",
-							Actions: []Action{
-								{},
-							},
-						},
-					},
-					CompletionType: CompletionTypeAllOf,
-					NumCompleted:   intstr.FromInt(1),
-				},
-			},
-			err: ``,
-		},
-		{
-			desp: "invalid completeType",
-			state: &State{
-				BaseState: BaseState{
-					Name: "1",
-					Type: "parallel",
-				},
-				ParallelState: &ParallelState{
-					Branches: []Branch{
-						{
-							Name: "b1",
-							Actions: []Action{
-								{},
-							},
-						},
-					},
-					CompletionType: CompletionTypeAllOf + "1",
-				},
-			},
-			err: `Key: 'State.ParallelState.CompletionType' Error:Field validation for 'CompletionType' failed on the 'oneof' tag`,
-		},
-		{
-			desp: "invalid numCompleted `int`",
-			state: &State{
-				BaseState: BaseState{
-					Name: "1",
-					Type: "parallel",
-				},
-				ParallelState: &ParallelState{
-					Branches: []Branch{
-						{
-							Name: "b1",
-							Actions: []Action{
-								{},
-							},
-						},
-					},
-					CompletionType: CompletionTypeAtLeast,
-					NumCompleted:   intstr.FromInt(0),
-				},
-			},
-			err: `Key: 'State.ParallelState.NumCompleted' Error:Field validation for 'NumCompleted' failed on the 'gt0' tag`,
-		},
-		{
-			desp: "invalid numCompleted string format",
-			state: &State{
-				BaseState: BaseState{
-					Name: "1",
-					Type: "parallel",
-				},
-				ParallelState: &ParallelState{
-					Branches: []Branch{
-						{
-							Name: "b1",
-							Actions: []Action{
-								{},
-							},
-						},
-					},
-					CompletionType: CompletionTypeAtLeast,
-					NumCompleted:   intstr.FromString("a"),
-				},
-			},
-			err: `Key: 'State.ParallelState.NumCompleted' Error:Field validation for 'NumCompleted' failed on the 'gt0' tag`,
-		},
-		{
-			desp: "normal",
-			state: &State{
-				BaseState: BaseState{
-					Name: "1",
-					Type: "parallel",
-				},
-				ParallelState: &ParallelState{
-					Branches: []Branch{
-						{
-							Name: "b1",
-							Actions: []Action{
-								{},
-							},
-						},
-					},
-					CompletionType: CompletionTypeAtLeast,
-					NumCompleted:   intstr.FromString("0"),
-				},
-			},
-			err: `Key: 'State.ParallelState.NumCompleted' Error:Field validation for 'NumCompleted' failed on the 'gt0' tag`,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.desp, func(t *testing.T) {
-			err := val.GetValidator().Struct(tc.state)
-
-			if tc.err != "" {
-				assert.Error(t, err)
-				assert.Regexp(t, tc.err, err)
-				return
-			}
-
-			assert.NoError(t, err)
 		})
 	}
 }
