+++ model/states.go
@@ -18,28 +18,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
-
-	validator "github.com/go-playground/validator/v10"
-	val "github.com/serverlessworkflow/sdk-go/v2/validator"
 )
-
-func init() {
-	val.GetValidator().RegisterStructValidation(stateStructLevelValidation, State{})
-}
-
-func stateStructLevelValidation(structLevel validator.StructLevel) {
-	state := structLevel.Current().Interface().(State)
-
-	hasTransition := state.Transition != nil
-	isEnd := state.End != nil && state.End.Terminate
-
-	// TODO: Improve message errors
-	if !hasTransition && !isEnd {
-		structLevel.ReportError(nil, "State.Transition,State.End", "transition,end", "required", "")
-	} else if hasTransition && isEnd {
-		structLevel.ReportError(nil, "State.Transition", "transition", "required_without", "end")
-	}
-}
 
 // StateType ...
 type StateType string
