+++ model/util_validator.go
@@ -1,0 +1,30 @@
+// Copyright 2020 The Serverless Workflow Specification Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package model
+
+import (
+	validator "github.com/go-playground/validator/v10"
+)
+
+func validTransitionAndEnd(structLevel validator.StructLevel, field interface{}, transition *Transition, end *End) {
+	hasTransition := transition != nil && transition.NextState != ""
+	isEnd := end != nil && (end.Terminate || end.ContinueAs != nil || len(end.ProduceEvents) > 0) // TODO: check the spec continueAs/produceEvents to see how it influences the end
+
+	if !hasTransition && !isEnd {
+		structLevel.ReportError(field, "Transition", "transition", "required", "must have one of transition, end")
+	} else if hasTransition && isEnd {
+		structLevel.ReportError(field, "Transition", "transition", "exclusive", "must have one of transition, end")
+	}
+}
