+++ parser/testdata/workflows/greetings_sleep.sw.json
@@ -20,7 +20,8 @@
         "timeouts": {
           "stateExecTimeout": "PT10S"
         },
-        "duration": "PT40S"
+        "duration": "PT40S",
+        "transition": "Greet"
       },
       {
         "name": "Greet",
